+++ src/core/eventEmitter.js
@@ -23,4 +23,4 @@ class EventEmitter {
   }
 }
 
-module.exports = EventEmitter
+export default new EventEmitter();